ftpgetinfo.c drop cast

+++ docs/examples/ftpgetinfo.c
@@ -42,7 +42,7 @@ static size_t throw_away(void *ptr, size_t size, size_t nmemb, void *data)
   (void)data;
   /* we are not interested in the headers itself,
      so we only return the size we would have saved ... */
-  return (size_t)(size * nmemb);
+  return size * nmemb;
 }
 
 int main(void)
